+++ packages/types/custom/core.ts
@@ -421,7 +421,7 @@ export interface Menu {
 
 export interface IconMenuConfiguration extends PluginOptions {
   menus: Menu[]
-  component?: VueConstructor
+  buttonComponent?: VueConstructor
   initiallyOpen?: string
 }