+++ packages/plugins/IconMenu/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## unpublished
 
-- Feature: Add new configuration parameter `component` to be able to interchange the UI component of this plugin.
+- Feature: Add new configuration parameter `buttonComponent` to be able to interchange the UI component of the used (icon) buttons.
+- Feature: Expose `IconMenuButton.vue` as a new export in order to use it in custom implementations.
 
 ## 1.3.1